pyproject.toml
[project]
name = "subtitle-convert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencc-python-reimplemented>=0.1.7",
]
